Set annotation color per threshold in 2D heatmaps
